Move runpaths.py to shared
It's no longer a cwrapper
