+++ src/ert/_c_wrappers/enkf/enkf_main.py
@@ -22,9 +22,9 @@ from ert._c_wrappers.enkf.enums import RealizationStateEnum
 from ert._c_wrappers.enkf.enums.ert_impl_type_enum import ErtImplType
 from ert._c_wrappers.enkf.ert_run_context import RunContext
 from ert._c_wrappers.enkf.model_config import ModelConfig
-from ert._c_wrappers.enkf.runpaths import Runpaths
 from ert._c_wrappers.job_queue import WorkflowRunner
 from ert._c_wrappers.util.substitution_list import SubstitutionList
+from ert.shared.runpaths import Runpaths
 
 if TYPE_CHECKING:
     from ert._c_wrappers.enkf import ErtConfig
